requirements.txt
anyio==3.6.2
asgi-lifespan==2.0.0
astroid==2.12.12
attrs==22.1.0
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.3
dill==0.3.6
dnspython==2.2.1
ecdsa==0.18.0
exceptiongroup==1.0.2
fastapi==0.85.2
h11==0.14.0
httptools==0.5.0
idna==3.4
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.8.0
mccabe==0.7.0
msal==1.20.0
packaging==21.3
passlib==1.7.4
platformdirs==2.5.2
pluggy==1.0.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
pylint==2.15.5
pymongo==4.3.2
pyparsing==3.0.9
pytest==7.2.0
pytest-asyncio==0.20.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
uvloop==0.17.0
watchfiles==0.18.0
websockets==10.4
wrapt==1.14.1
